refactor: delete safeguard api update (#10992)

+++ frontend/src/component/feature/FeatureView/FeatureOverview/ReleasePlan/ReleasePlan.tsx
@@ -409,11 +409,13 @@ export const ReleasePlan = ({
 
     const handleSafeguardDelete = async () => {
         try {
+            if (safeguards.length === 0) return;
             await deleteSafeguard({
                 projectId,
                 featureName,
                 environment,
                 planId: id,
+                safeguardId: safeguards[0].id,
             });
             setToastData({
                 type: 'success',
